# Improve performance of formatting (both with and w/o `color` enabled).