--- a/Cargo.toml
+++ b/Cargo.toml
@@ -18,7 +18,7 @@ A simple to use, efficient, and full featured  Command Line Argument Parser
 bitflags  = "~0.7"
 vec_map   = "~0.6"
 libc      = { version = "~0.2.9",  optional = true }
-ansi_term = { version = "~0.9.0",  optional = true }
+termion   = { version = "~1.1.1",  optional = true }
 strsim    = { version = "~0.5.1",  optional = true }
 yaml-rust = { version = "~0.3.2",  optional = true }
 clippy    = { version = "~0.0.88", optional = true }
@@ -32,7 +32,7 @@ regex = "~0.1.69"
 [features]
 default     = ["suggestions", "color", "wrap_help"]
 suggestions = ["strsim"]
-color       = ["ansi_term", "libc"]
+color       = ["termion", "libc"]
 yaml        = ["yaml-rust"]
 wrap_help   = ["libc", "term_size"]
 lints       = ["clippy", "nightly"]