--- a/Cargo.toml
+++ b/Cargo.toml
@@ -18,13 +18,13 @@ A simple to use, efficient, and full featured  Command Line Argument Parser
 bitflags  = "~0.7"
 vec_map   = "~0.6"
 libc      = { version = "~0.2.9",  optional = true }
-termion   = { version = "~1.1.1",  optional = true }
 strsim    = { version = "~0.5.1",  optional = true }
 yaml-rust = { version = "~0.3.2",  optional = true }
 clippy    = { version = "~0.0.88", optional = true }
 unicode-width = "~0.1.3"
 unicode-segmentation = "~0.1.2"
 term_size = { version = "~0.2.0", optional = true }
+termion   = { version = "~1.1.1", optional = true }
 
 [dev-dependencies]
 regex = "~0.1.69"