--- a/src/fmt.rs
+++ b/src/fmt.rs
@@ -1,8 +1,5 @@
 #[cfg(all(feature = "color", not(target_os = "windows")))]
-use ansi_term::ANSIString;
-
-#[cfg(all(feature = "color", not(target_os = "windows")))]
-use ansi_term::Colour::{Green, Red, Yellow};
+use termion::{color, style};
 
 #[cfg(feature = "color")]
 use libc;
@@ -52,10 +49,10 @@ macro_rules! color {
             ColorWhen::Auto => if is_a_tty($_self.use_stderr) {
                 Format::$c($m)
             } else {
-                Format::None($m)
+                Format::none($m)
             },
             ColorWhen::Always => Format::$c($m),
-            ColorWhen::Never => Format::None($m),
+            ColorWhen::Never => Format::none($m),
         }
     };
 }
@@ -65,28 +62,28 @@ impl Colorizer {
         where T: fmt::Display + AsRef<str>
     {
         debugln!("fn=good;");
-        color!(self, Good, msg)
+        color!(self, good, msg)
     }
 
     pub fn warning<T>(&self, msg: T) -> Format<T>
         where T: fmt::Display + AsRef<str>
     {
         debugln!("fn=warning;");
-        color!(self, Warning, msg)
+        color!(self, warning, msg)
     }
 
     pub fn error<T>(&self, msg: T) -> Format<T>
         where T: fmt::Display + AsRef<str>
     {
         debugln!("fn=error;");
-        color!(self, Error, msg)
+        color!(self, error, msg)
     }
 
     pub fn none<T>(&self, msg: T) -> Format<T>
         where T: fmt::Display + AsRef<str>
     {
         debugln!("fn=none;");
-        Format::None(msg)
+        Format::none(msg)
     }
 }
 
@@ -101,75 +98,120 @@ impl Default for Colorizer {
 
 /// Defines styles for different types of error messages. Defaults to Error=Red, Warning=Yellow,
 /// and Good=Green
-#[derive(Debug)]
-#[doc(hidden)]
-pub enum Format<T> {
+#[derive(Debug, Copy, Clone)]
+pub enum Kind {
     /// Defines the style used for errors, defaults to Red
-    Error(T),
+    Error,
     /// Defines the style used for warnings, defaults to Yellow
-    Warning(T),
+    Warning,
     /// Defines the style used for good values, defaults to Green
-    Good(T),
+    Good,
     /// Defines no formatting style
-    None(T),
+    None,
 }
 
-#[cfg(all(feature = "color", not(target_os = "windows")))]
-impl<T: AsRef<str>> Format<T> {
-    fn format(&self) -> ANSIString {
-        match *self {
-            Format::Error(ref e) => Red.bold().paint(e.as_ref()),
-            Format::Warning(ref e) => Yellow.paint(e.as_ref()),
-            Format::Good(ref e) => Green.paint(e.as_ref()),
-            Format::None(ref e) => ANSIString::from(e.as_ref()),
+/// An log message.
+#[doc(hidden)]
+#[derive(Debug)]
+pub struct Format<T> {
+    kind: Kind,
+    msg: T,
+}
+
+impl<T> Format<T> {
+    /// An error message.
+    pub fn error(msg: T) -> Format<T> {
+        Format {
+            kind: Kind::Error,
+            msg: msg,
+        }
+    }
+
+    /// A warning message.
+    pub fn warning(msg: T) -> Format<T> {
+        Format {
+            kind: Kind::Warning,
+            msg: msg,
+        }
+    }
+
+    /// A "good" message.
+    ///
+    /// That is, it went well.
+    pub fn good(msg: T) -> Format<T> {
+        Format {
+            kind: Kind::Good,
+            msg: msg,
+        }
+    }
+
+    /// A "neutral" message.
+    ///
+    /// That is, it is simply a log message.
+    pub fn none(msg: T) -> Format<T> {
+        Format {
+            kind: Kind::None,
+            msg: msg,
         }
     }
 }
 
-#[cfg(any(not(feature = "color"), target_os = "windows"))]
-impl<T: fmt::Display> Format<T> {
-    fn format(&self) -> &T {
-        match *self {
-            Format::Error(ref e) => e,
-            Format::Warning(ref e) => e,
-            Format::Good(ref e) => e,
-            Format::None(ref e) => e,
-        }
-    }
-}
-
-
 #[cfg(all(feature = "color", not(target_os = "windows")))]
-impl<T: AsRef<str>> fmt::Display for Format<T> {
-    fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
-        write!(f, "{}", &self.format())
-    }
-}
-
-#[cfg(any(not(feature = "color"), target_os = "windows"))]
 impl<T: fmt::Display> fmt::Display for Format<T> {
     fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
-        write!(f, "{}", &self.format())
+        match self.kind {
+            Kind::Error => write!(f, "{color}{style}{msg}{reset}",
+                                    color = color::Fg(color::Red),
+                                    style = style::Bold,
+                                    msg   = self.msg,
+                                    reset = style::Reset),
+            Kind::Warning => write!(f, "{color}{msg}{reset}",
+                                      color = color::Fg(color::Yellow),
+                                      msg   = self.msg,
+                                      reset = style::Reset),
+            Kind::Good => write!(f, "{color}{msg}{reset}",
+                                   color = color::Fg(color::Green),
+                                   msg   = self.msg,
+                                   reset = style::Reset),
+            Kind::None => write!(f, "{}", self.msg),
+        }
+    }
+}
+
+#[cfg(any(not(feature = "color"), target_os = "windows"))]
+impl<T: AsRef<str>> fmt::Display for Format<T> {
+    fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
+        write!(f, self.msg.as_ref())
     }
 }
 
 #[cfg(all(test, feature = "color", not(target_os = "windows")))]
 mod test {
-    use ansi_term::ANSIString;
-    use ansi_term::Colour::{Green, Red, Yellow};
-    use super::Format;
+    use termion::{color, style};
+    use super::*;
 
     #[test]
     fn colored_output() {
-        let err = Format::Error("error");
-        assert_eq!(&*format!("{}", err),
-                   &*format!("{}", Red.bold().paint("error")));
-        let good = Format::Good("good");
-        assert_eq!(&*format!("{}", good), &*format!("{}", Green.paint("good")));
-        let warn = Format::Warning("warn");
-        assert_eq!(&*format!("{}", warn), &*format!("{}", Yellow.paint("warn")));
-        let none = Format::None("none");
+        let err = Format {
+            kind: Kind::Error,
+            msg: "error",
+        };
+        assert_eq!(&*format!("{}", err), &*format!("{}{}error{}", color::Fg(color::Red), style::Bold, style::Reset));
+        let good = Format {
+            kind: Kind::Good,
+            msg: "good",
+        };
+        assert_eq!(&*format!("{}", good), &*format!("{}good{}", color::Fg(color::Green), style::Reset));
+        let warn = Format {
+            kind: Kind::Warning,
+            msg: "warn",
+        };
+        assert_eq!(&*format!("{}", warn), &*format!("{}warn{}", color::Fg(color::Yellow), style::Reset));
+        let none = Format {
+            kind: Kind::None,
+            msg: "none",
+        };
         assert_eq!(&*format!("{}", none),
-                   &*format!("{}", ANSIString::from("none")));
+                   "none");
     }
 }