Instead of using `ansi_term`, we change clap to use `termion`, which
does not encode colors as an enum, but instead as distinct types
implementing `Display` (and hence avoids branches).

Old:

```
running 10 tests
test example1          ... bench:      20,448 ns/iter (+/- 384)
test example10         ... bench:       7,646 ns/iter (+/- 2,363)
test example2          ... bench:         811 ns/iter (+/- 18)
test example3          ... bench:      18,206 ns/iter (+/- 602)
test example4          ... bench:       9,876 ns/iter (+/- 5,241)
test example4_template ... bench:      10,651 ns/iter (+/- 7,594)
test example5          ... bench:       6,994 ns/iter (+/- 1,411)
test example6          ... bench:       3,542 ns/iter (+/- 3,261)
test example7          ... bench:      13,315 ns/iter (+/- 6,517)
test example8          ... bench:       8,056 ns/iter (+/- 394)
```

New:

```
running 10 tests
test example1          ... bench:      19,681 ns/iter (+/- 3,156)
test example10         ... bench:       7,147 ns/iter (+/- 2,121)
test example2          ... bench:         825 ns/iter (+/- 76)
test example3          ... bench:      17,180 ns/iter (+/- 1,065)
test example4          ... bench:       9,075 ns/iter (+/- 993)
test example4_template ... bench:       9,828 ns/iter (+/- 266)
test example5          ... bench:       6,336 ns/iter (+/- 723)
test example6          ... bench:       3,485 ns/iter (+/- 217)
test example7          ... bench:       7,626 ns/iter (+/- 375)
test example8          ... bench:       7,592 ns/iter (+/- 396)
```

(both compiled with --opt-level=3 and panic=abort)

Further optimization can perhaps be done by eliminating the virtual
calls in `fmt::Formatter`, which, for some strange reason, holds a
`fmt::Write` trait object. Fortunately, most of it is eliminated in the
LLVM devirtualization pass.