@kbknapp 2016-09-28
Ah wow, this'll be great! I've only briefly looked at termion and seen it mentioned a few times in reddit, so this is very cool. Once the CI is happy this should be a great merge :smile: