@kbknapp 2016-10-04
@ticki still getting an error for Windows. But I'm also waiting on #674 and #670 to merge too, so no rush since this'll require a rebase to merge anyways. :+1: