@homu 2016-10-05
:umbrella: The latest upstream changes (presumably #681) made this pull request unmergeable. Please resolve the merge conflicts.