@BurntSushi 2017-01-17
Last time I tried termion, I found that it didn't compile on Windows, which means you'll need to make it a Unix-only dep. On top of that, native Windows consoles don't support ANSI escape sequences. To support colors on Windows, you'll either need to use the `term` or `termcolor` crates.