@kbknapp 2017-01-17
@BurntSushi yeah this (output coloring) is on my list of things to at least look at and *possibly* overhaul at some point. If `termion` turns out not to be the one, I'm sort of indifferent to which one it ends up being so long as it does what I need it to do [well], and is stable/supported. I'd prefer consistency wherever possible though and if I could ever get to 100% functionality on all platforms that'd be best :smile: 

In fact I'll just close this PR until I can either look into the details again and determine if I want to support it, pick something else, or just stick with what we've got.