@BurntSushi 2017-01-17
@kbknapp I've spent an embarrassing amount of time on getting colors working correctly across Windows, Linux and Mac for ripgrep, *including* mintty/MSYS2 and native Windows console support. If you want to save yourself a lot of time, just do what ripgrep does. ;-)