@kbknapp 2017-01-18
That's actually the plan 😄 I've been spending some time lately going through (read admiring) your work on ripgrep, and it's even becoming somewhat of a routine for me:

1. Need to do new thing...
2. *check @BurntSushi's work guidance*
3. Do thing